tavern_update_v3
updated tavern with jukebox v3
