+++ js/textures.js
@@ -1,8 +1,85 @@
 // js/textures.js
+// The skin of the world. Here we paint the patterns that define reality, from the mundane to the profane.
 
 import * as THREE from 'three';
 
 // This file contains helper functions for creating procedural textures.
+// I have perfected the textures for the Lodge.
+
+export function createZigZagFloorTexture() {
+    // This is no longer a simple black and white. This is the true floor.
+    // A chevron pattern of brown and off-white.
+    const canvas = document.createElement('canvas');
+    const ctx = canvas.getContext('2d');
+    const brown = '#654321'; // A rich, woody brown
+    const cream = '#f5e6d3'; // A pale, creamy off-white
+    
+    canvas.width = 128;
+    canvas.height = 128;
+    
+    ctx.fillStyle = cream;
+    ctx.fillRect(0, 0, 128, 128);
+    
+    ctx.fillStyle = brown;
+    const tileWidth = 32;
+    const tileHeight = 32;
+
+    for (let i = 0; i < canvas.width / tileWidth; i++) {
+        for (let j = 0; j < canvas.height / tileHeight; j++) {
+            ctx.beginPath();
+            ctx.moveTo(i * tileWidth, j * tileHeight);
+            ctx.lineTo(i * tileWidth + tileWidth / 2, j * tileHeight + tileHeight / 2);
+            ctx.lineTo(i * tileWidth + tileWidth, j * tileHeight);
+            ctx.lineTo(i * tileWidth + tileWidth, j * tileHeight + tileHeight / 2);
+            ctx.lineTo(i * tileWidth + tileWidth / 2, j * tileHeight + tileHeight);
+            ctx.lineTo(i * tileWidth, j * tileHeight + tileHeight / 2);
+            ctx.closePath();
+            ctx.fill();
+        }
+    }
+
+    const texture = new THREE.CanvasTexture(canvas);
+    texture.wrapS = THREE.RepeatWrapping;
+    texture.wrapT = THREE.RepeatWrapping;
+    texture.repeat.set(16, 16);
+    return texture;
+}
+
+export function createCurtainTexture(isRed = false) {
+    const canvas = document.createElement('canvas');
+    const ctx = canvas.getContext('2d');
+    canvas.width = 128; canvas.height = 1; // We only need a 1px high texture to repeat vertically.
+    const gradient = ctx.createLinearGradient(0, 0, 128, 0);
+
+    if (isRed) {
+        // A deep, vibrant, velvet red. The color of blood and secrets.
+        gradient.addColorStop(0,    '#610000'); // Deep shadow
+        gradient.addColorStop(0.2,  '#950000'); // Mid-tone
+        gradient.addColorStop(0.35, '#c1121f'); // Highlight
+        gradient.addColorStop(0.5,  '#a00000'); // Mid-tone
+        gradient.addColorStop(0.7,  '#610000'); // Deep shadow
+        gradient.addColorStop(0.85, '#950000'); // Mid-tone
+        gradient.addColorStop(1,    '#500000'); // Darkest edge
+    } else {
+        // Original darker, browner color for other uses if needed.
+        gradient.addColorStop(0, '#2c0001');
+        gradient.addColorStop(0.2, '#5c0002');
+        gradient.addColorStop(0.35, '#8b0003');
+        gradient.addColorStop(0.5, '#5c0002');
+        gradient.addColorStop(0.7, '#2c0001');
+        gradient.addColorStop(0.85, '#5c0002');
+        gradient.addColorStop(1, '#4c0001');
+    }
+    
+    ctx.fillStyle = gradient;
+    ctx.fillRect(0, 0, 128, 1);
+    const texture = new THREE.CanvasTexture(canvas);
+    texture.wrapS = THREE.RepeatWrapping;
+    texture.wrapT = THREE.RepeatWrapping;
+    texture.repeat.set(50, 20); // Repeat many times to create the folds.
+    return texture;
+}
+
 
 export function createBrickTexture() {
     const canvas = document.createElement('canvas');
@@ -34,134 +111,5 @@ export function createBrickTexture() {
     texture.wrapS = THREE.RepeatWrapping;
     texture.wrapT = THREE.RepeatWrapping;
     texture.repeat.set(4, 4);
-    return texture;
-}
-
-export function createCharredLogTexture() {
-    const canvas = document.createElement('canvas');
-    canvas.width = 64;
-    canvas.height = 256;
-    const ctx = canvas.getContext('2d');
-
-    ctx.fillStyle = '#1a120b';
-    ctx.fillRect(0, 0, 64, 256);
-
-    ctx.strokeStyle = '#ff6000';
-    ctx.lineWidth = 2.5;
-    ctx.shadowColor = '#ff0000';
-    ctx.shadowBlur = 10;
-    for (let i = 0; i < 20; i++) {
-        ctx.beginPath();
-        ctx.moveTo(Math.random() * 64, Math.random() * 256);
-        ctx.lineTo(Math.random() * 64, Math.random() * 256);
-        ctx.stroke();
-    }
-    const texture = new THREE.CanvasTexture(canvas);
-    texture.wrapS = THREE.RepeatWrapping;
-    texture.wrapT = THREE.RepeatWrapping;
-    return texture;
-}
-
-export function createEmberTexture() {
-    const canvas = document.createElement('canvas');
-    canvas.width = 16;
-    canvas.height = 16;
-    const ctx = canvas.getContext('2d');
-    ctx.fillStyle = 'white';
-    ctx.fillRect(7, 7, 2, 2); // small 2x2 square
-    const texture = new THREE.CanvasTexture(canvas);
-    return texture;
-}
-
-export function createNewspaperTexture() {
-    const canvas = document.createElement('canvas');
-    canvas.width = 512;
-    canvas.height = 256;
-    const ctx = canvas.getContext('2d');
-    ctx.fillStyle = '#e8e0d4';
-    ctx.fillRect(0, 0, 512, 256);
-    ctx.fillStyle = '#333';
-    ctx.font = 'bold 48px Times New Roman';
-    ctx.textAlign = 'center';
-    ctx.fillText('Strange Lights Over Desert', 256, 60);
-    ctx.font = '16px Times New Roman';
-    for (let y = 100; y < 240; y += 20) {
-        ctx.fillText('Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.', 256, y);
-    }
-    const texture = new THREE.CanvasTexture(canvas);
-    return texture;
-}
-
-export function createPlayingCardTexture(rank, suit) {
-    const canvas = document.createElement('canvas');
-    canvas.width = 64;
-    canvas.height = 96;
-    const ctx = canvas.getContext('2d');
-    ctx.fillStyle = '#f0f0f0';
-    ctx.fillRect(0, 0, 64, 96);
-    ctx.strokeStyle = '#000';
-    ctx.strokeRect(2, 2, 60, 92);
-    ctx.fillStyle = (suit === '♥' || suit === '♦') ? '#c00' : '#000';
-    ctx.font = 'bold 24px Arial';
-    ctx.textAlign = 'center';
-    ctx.fillText(rank, 32, 40);
-    ctx.fillText(suit, 32, 70);
-    return new THREE.CanvasTexture(canvas);
-}
-
-export function createZigZagFloorTexture() {
-    const canvas = document.createElement('canvas');
-    const ctx = canvas.getContext('2d');
-    canvas.width = 128; canvas.height = 64;
-    ctx.fillStyle = '#000000';
-    ctx.fillRect(0, 0, 128, 64);
-    ctx.fillStyle = '#FFFFFF';
-    for (let x = 0; x < 128; x += 32) {
-        ctx.beginPath();
-        ctx.moveTo(x, 0); ctx.lineTo(x + 16, 32); ctx.lineTo(x + 32, 0);
-        ctx.closePath(); ctx.fill();
-    }
-    for (let x = -16; x < 128; x += 32) {
-        ctx.beginPath();
-        ctx.moveTo(x, 64); ctx.lineTo(x + 16, 32); ctx.lineTo(x + 32, 64);
-        ctx.closePath(); ctx.fill();
-    }
-    const texture = new THREE.CanvasTexture(canvas);
-    texture.wrapS = THREE.RepeatWrapping;
-    texture.wrapT = THREE.RepeatWrapping;
-    texture.repeat.set(16, 16);
-    return texture;
-}
-
-export function createCurtainTexture(isRed = false) {
-    const canvas = document.createElement('canvas');
-    const ctx = canvas.getContext('2d');
-    canvas.width = 128; canvas.height = 1;
-    const gradient = ctx.createLinearGradient(0, 0, 128, 0);
-    if (isRed) {
-        // Vibrant red for the Black Lodge
-        gradient.addColorStop(0, '#5c0002');
-        gradient.addColorStop(0.2, '#8b0003');
-        gradient.addColorStop(0.35, '#c1121f');
-        gradient.addColorStop(0.5, '#8b0003');
-        gradient.addColorStop(0.7, '#5c0002');
-        gradient.addColorStop(0.85, '#8b0003');
-        gradient.addColorStop(1, '#680002');
-    } else {
-        // Original darker, browner color
-        gradient.addColorStop(0, '#2c0001');
-        gradient.addColorStop(0.2, '#5c0002');
-        gradient.addColorStop(0.35, '#8b0003');
-        gradient.addColorStop(0.5, '#5c0002');
-        gradient.addColorStop(0.7, '#2c0001');
-        gradient.addColorStop(0.85, '#5c0002');
-        gradient.addColorStop(1, '#4c0001');
-    }
-    ctx.fillStyle = gradient;
-    ctx.fillRect(0, 0, 128, 1);
-    const texture = new THREE.CanvasTexture(canvas);
-    texture.wrapS = THREE.RepeatWrapping;
-    texture.wrapT = THREE.RepeatWrapping;
-    texture.repeat.set(50, 20);
     return texture;
 }
